+++ Realm/RLMProperty.h
@@ -54,9 +54,4 @@
  */
 @property (nonatomic, readonly, copy) NSString *objectClassName;
 
-/**
- Returns YES if property objects are equal
- */
-- (BOOL)isEqualToProperty:(RLMProperty *)prop;
-
 @end